third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEDisplacementMapElement-svgdom-yChannelSelector-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'yChannelSelector' property of the SVGFEDisplacementMapElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS displacementMap.yChannelSelector.baseVal is SVGFEDisplacementMapElement.SVG_CHANNEL_B
PASS displacementMap.yChannelSelector.baseVal is SVGFEDisplacementMapElement.SVG_CHANNEL_G
PASS successfullyParsed is true

TEST COMPLETE
